Credential check prior to issuing session (#1)
* Authentication Process via POST Request

A POST request will contain the JSON credentials of the user.  This will be compared to the backend and a true or false will be returned to confirm whether or not the credentials are valid.

* Decode JSON from HTTP Request Body

JSON is now decoded properly.

* Query row in database to find user

User can be found in the database.  Will return false and throw error if user is not present.

* Get user password from database

User password is now returned from database if the user is present.

* Changed Error output

* Session created if password challenge passed

If a user succeeds in the password challenge then a session is created.   Otherwise a 401 status is returned.